+++ apps/block_scout_web/lib/block_scout_web/views/address_contract_verification_via_sourcify_view.ex
@@ -1,0 +1,3 @@
+defmodule BlockScoutWeb.AddressContractVerificationViaSourcifyView do
+  use BlockScoutWeb, :view
+end