+++ apps/block_scout_web/lib/block_scout_web/web_router.ex
@@ -157,6 +157,13 @@ defmodule BlockScoutWeb.WebRouter do
       )
 
       resources(
+        "/verify-via-sourcify",
+        AddressContractVerificationViaSourcifyController,
+        only: [:new],
+        as: :verify_contract_via_sourcify
+      )
+
+      resources(
         "/read-contract",
         AddressReadContractController,
         only: [:index, :show],